% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\encoding{UTF-8}
\name{get_dataset}
\alias{get_dataset}
\title{Fonction permettant d'importer les données de l'API 'données locales' pour un couple de paramètre}
\usage{
get_dataset(
  jeu_donnees,
  croisement,
  modalite,
  nivgeo,
  codgeo,
  temporisation = NA
)
}
\arguments{
\item{jeu_donnees}{code jeu de données se composant du nom de la source, du millésime des données et parfois du millésime géographique de diffusion}

\item{croisement}{sélection de variables (composée d'une variable ou de plusieurs)}

\item{modalite}{modalités souhaitées pour les variables (dans le même ordre)}

\item{nivgeo}{niveau géographique du code demandé}

\item{codgeo}{Codes géographiques souhaité}

\item{temporisation}{temps d'attente entre chaque requête (si utilisé dans une boucle)}
}
\value{
une liste contenant 4 data.frame. un contenant les données, métadonnées, les infos sur la zone sélectionnées et les infos sur la source
}
\description{
Fonction permettant d'importer les données de l'API 'données locales' pour un couple de paramètre
}
\examples{
\donttest{
  get_dataset("GEO2017REE2017",
              "NA5_B-ENTR_INDIVIDUELLE",
              "all.all",
              "COM",
              "51108")
}
}
